[ALPS03237779] memcfg: add reserve_memory and refactor code
[Detail]
Add reserve_memory and refactor memory_layout/total_reserve.
Move reserved memory related functions to mtk_memcfg_reserve_info.c.

MTK-Commit-Id: 4abac7d32a4dae3e65d2705d12056ab47dd65a87

CR-Id: ALPS03237779
Feature: 4GB_DRAM

+++ include/linux/mm.h
@@ -31,16 +31,16 @@ struct writeback_control;
 struct bdi_writeback;
 
 struct kernel_reserve_meminfo {
-	unsigned long total;
-	unsigned long available;
-	unsigned long kernel_code;
-	unsigned long rwdata;
-	unsigned long rodata;
-	unsigned long init;
-	unsigned long bss;
-	unsigned long reserved;
+	unsigned long long total;
+	unsigned long long available;
+	unsigned long long kernel_code;
+	unsigned long long rwdata;
+	unsigned long long rodata;
+	unsigned long long init;
+	unsigned long long bss;
+	unsigned long long reserved;
 #ifdef CONFIG_HIGHMEM
-	unsigned long highmem;
+	unsigned long long highmem;
 #endif
 };
